share hardware lib
